refactor(server-core): parameterize the static console serving and the console login
defineStaticConsole gives every static console its own dist and html cache
(two consoles sharing a module-level slot would serve one bundle's shell for
the other), and ConsoleLogin lifts the plan-088 kick and redemption out of
AccountController, parameterized by client name and path segment. No
behaviour change: the account console is the only instance so far.
