+++ apps/server-core/src/adapters/http/ui/index.ts
@@ -9,4 +9,5 @@ export * from './account-console/index.ts';
 export * from './auth-console/index.ts';
 export * from './console-packages/index.ts';
 export * from './shared/index.ts';
+export * from './static-console/index.ts';
 export * from './theme/index.ts';
